requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
anthropic==0.76.0
anyio==4.12.1
attrs==25.4.0
babel==2.17.0
blinker==1.9.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
courlan==1.3.2
dateparser==1.2.2
defusedxml==0.7.1
distro==1.9.0
docstring_parser==0.17.0
edge-tts==7.2.7
Flask==3.1.2
frozenlist==1.8.0
google-api-core==2.29.0
google-api-python-client==2.188.0
google-auth==2.47.0
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.2.4
googleapis-common-protos==1.72.0
h11==0.16.0
htmldate==1.9.4
httpcore==1.0.9
httplib2==0.31.1
httpx==0.28.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.12.0
jusText==3.0.2
lxml==6.0.2
lxml_html_clean==0.4.3
MarkupSafe==3.0.3
multidict==6.7.1
oauthlib==3.3.1
propcache==0.4.1
proto-plus==1.27.0
protobuf==6.33.4
pyasn1==0.6.2
pyasn1_modules==0.4.2
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
regex==2026.1.15
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
tabulate==0.9.0
tld==0.13.1
trafilatura==2.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.3
Werkzeug==3.1.5
yarl==1.22.0
youtube-transcript-api==1.2.3
yt-dlp==2025.12.8
